+++ src/web/UrlManager.php
@@ -142,7 +142,7 @@ class UrlManager extends \yii\web\UrlManager
         $route = trim($params[0], '/');
         unset($params[0]);
 
-        return UrlHelper::actionUrl($route, $params);
+        return UrlHelper::actionUrl($route, $params, null, false);
     }
 
     /**